refactor and change to safeparse
