+++ site/schemas/hootsuite.schema.ts
@@ -5,6 +5,8 @@ export const hootsuiteTokenSchema = z.object({
     access_token: z.string(),
 });
 
+export type HootsuiteToken = z.infer<typeof hootsuiteMeSchema>;
+
 export const hootsuiteMeSchema = z.object({
     data: z.object({
         id: z.string(),
